Add custom types support
Removes &to_string/1 in &number_to_delimited/1 because it's a reason for
a failure of custom types what implement `Number.Conversion` conversion
protocol but do not implement `String.Chars` (or implements it not the
format that Decimal expects). And simply there is no need in it at all.

But the &number_to_phone/1 kept untouched because it heavily relies on
string representation and if someone introduce a special type for
the phone number representation it's doubtful that he would use any other
functions from the package for that type.